+++ CHANGELOG.md
@@ -129,3 +129,7 @@ Added Calculator GUI fix for "IC2 crop breeding plugin"
 ## 1.6.0
 Added Fix for vanilla falling block dupe glitch  
 Added coremod and early mixin support  
+## 1.6.1
+Added Fix for storage drawers extras:  
+- Extra trim subtypes, which fixes silk touch drop of it  
+- Localization key fix